requirements.txt
pandas==1.1.0
numpy==1.19.1
xgboost==1.1.1
scikit-learn==0.23.1
dask[dataframe]==0.19.4
lunardate==0.2.0
convertdate==2.2.1
matplotlib==3.2.1
fbprophet==0.6
python-dateutil==2.8.1
